Update 06/03/2022
 create chart

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Category;
+use Illuminate\Support\Facades\DB;
+use App\Models\Idea;
 
 class CategoryController extends Controller
 {
@@ -57,9 +59,21 @@ class CategoryController extends Controller
          */
         public function destroy(Category $category)
         {
-            $category->delete();
-    
-            return redirect('/category')->with('success','Category deleted successfully');
+            $category1 = DB::table('categories')
+            ->where('categories.id', $category->id)
+            ->join('ideas', 'ideas.category_id', '=', 'categories.id')
+            ->select(array('categories.*', DB::raw('count(ideas.id) as ideas_count')))
+            ->groupBy('ideas.category_id')
+            ->get();
+     
+            if (empty($category1[0])) {
+                $category->delete();
+                return redirect('/category')->with('success','Category deleted successfully');
+            }
+            elseif ($category1[0]->ideas_count > 0) {
+                return redirect('/category')->with('error','Category cannot be deleted because it has ideas');
+            }
+
         }
 }
